requirements.txt
pybind11
ninja
wheel
numpy<=1.26.0
six
regex
decorator
attrs
psutil
pyyaml
protobuf
einops
scipy
sentencepiece
pytest
pytest-mock
transformers>=4.43.2
gpytorch
pandas
scikit-learn
SQLAlchemy
